wip optimize types
