+++ console/packages/starwhale-ui/src/Upload/types.ts
@@ -1,0 +1,5 @@
+export type UploadFile = File & {
+    path: string
+    originFileObj?: any
+    status: 'uploading' | 'error' | 'done' | 'error_max' | 'error_exist'
+}
